test: Split models.detection unittests
